#82 correct minor problem in unithelpers

+++ module/Administrator/test/AdministratorTest/Controller/UnitHelpers.php
@@ -392,7 +392,7 @@ abstract class UnitHelpers extends \PHPUnit_Framework_TestCase
         return $repository->Create([
                     'duedate' => new \DateTime,
                     'description' => uniqid() . ' Description for independentwork',
-                    'durationAK' => uniqid(),
+                    'durationAK' => (int) uniqid(),
                     'subjectRound' => $subjectRound->getId(),
                     'teacher' => $teacher->getId(),
         ]);
